Big clean up and bugfixing
