+++ rbgh/package.json
@@ -1,6 +1,6 @@
 {
   "name": "rbgh",
-  "version": "0.1.1",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.4",
