+++ rbgh/src/App.tsx
@@ -1,93 +1,36 @@
-import React, { Reducer, useReducer } from 'react'
-import { BrowserRouter as Router } from 'react-router-dom'
+import './css/App.css' // Styles
+import { allMethods } from './data'
 import { AppRoutes } from './Routes'
+import { BrowserRouter as Router } from 'react-router-dom'
 import { Header } from './components/Header'
-import { Player } from './types'
+import { MethodOption, Player, PlayerAction } from './types'
+import { playerReducer } from './Store'
 import RBGHContext from './RBGHContext'
-// Styles
-import './css/App.css'
+import React, { Reducer, useReducer, useState } from 'react'
+
+export const DEVELOPMENT_MODE = process.env.NODE_ENV === 'development'
+export const MIN_PLAYABLE_PLAYERS = 3
+export const MAX_PLAYERS = 6
 
 const initialPlayerList = [
   {
-    name: 'RenameMe',
+    name: '',
     id: 1,
     show: true,
   },
 ]
 
-export interface PlayerAction {
-  type: ValidPlayerReducerActionTypes
-  payload: { id: number; name: string }
-}
-
-export enum PlayerReducerActionTypes {
-  ADD_PLAYER = 'ADD_PLAYER',
-  HIDE_PLAYER = 'HIDE_PLAYER',
-  UPDATE_PLAYER = 'UPDATE_PLAYER',
-  REMOVE_PLAYER = 'REMOVE_PLAYER',
-}
-
-export type ValidPlayerReducerActionTypes =
-  | PlayerReducerActionTypes.ADD_PLAYER
-  | PlayerReducerActionTypes.HIDE_PLAYER
-  | PlayerReducerActionTypes.REMOVE_PLAYER
-  | PlayerReducerActionTypes.UPDATE_PLAYER
-
-const playerReducer = (state: Player[], action: PlayerAction): Player[] => {
-  let id: number
-  let index: number
-  const MAX_PLAYERS = 6
-
-  switch (action.type) {
-    case PlayerReducerActionTypes.ADD_PLAYER:
-      if (state.length < MAX_PLAYERS) {
-        console.log(`Running playerReducer, final state is:`, { state, action })
-        return [...state, { name: '', id: new Date().getTime(), show: true }]
-      }
-      return state
-    case PlayerReducerActionTypes.HIDE_PLAYER:
-      id = action.payload?.id
-      index = state.findIndex((player: Player) => player.id === id)
-      if (id && index !== -1 && state[index].show) {
-        state[index].show = false
-        console.log(`Running playerReducer, final state is:`, { state, action })
-        return [...state]
-      }
-      return state
-    case PlayerReducerActionTypes.UPDATE_PLAYER:
-      const newName = action.payload?.name
-      id = action.payload.id
-      index = state.findIndex((player: Player) => player.id === id)
-      if (id && index !== -1) {
-        let player = {
-          ...state[index],
-          name: newName,
-        }
-        state[index] = player
-        console.log(`Running playerReducer, final state is:`, { state, action })
-        return [...state]
-      }
-      return state
-    case PlayerReducerActionTypes.REMOVE_PLAYER:
-      id = action.payload?.id
-      if (id) {
-        console.log(`Running playerReducer, final state is:`, { state, action })
-        return state.filter((player: Player) => player.id !== id)
-      }
-      return state
-    default:
-      throw new Error('unknown action!')
-  }
-}
-
 function App() {
   const [playerList, playerDispatch] = useReducer<Reducer<Player[], PlayerAction>>(playerReducer, [
     ...initialPlayerList,
   ])
+  const [methodList, setMethodList] = useState<MethodOption[]>([...allMethods])
 
   const passProps = {
     playerList,
     playerDispatch,
+    methodList,
+    setMethodList,
   }
 
   return (
@@ -97,6 +40,12 @@ function App() {
           <Header />
           <main className="content">
             <AppRoutes />
+
+            {DEVELOPMENT_MODE && (
+              <small>
+                <pre>{JSON.stringify({ playerList }, null, 2)}</pre>
+              </small>
+            )}
           </main>
         </Router>
       </div>
