+++ rbgh/src/types/Reducer.ts
@@ -1,0 +1,17 @@
+export enum PlayerReducerActionTypes {
+  ADD_PLAYER = 'ADD_PLAYER',
+  HIDE_PLAYER = 'HIDE_PLAYER',
+  REMOVE_PLAYER = 'REMOVE_PLAYER',
+  UPDATE_PLAYER = 'UPDATE_PLAYER',
+}
+
+export type ValidPlayerReducerActionTypes =
+  | PlayerReducerActionTypes.ADD_PLAYER
+  | PlayerReducerActionTypes.HIDE_PLAYER
+  | PlayerReducerActionTypes.REMOVE_PLAYER
+  | PlayerReducerActionTypes.UPDATE_PLAYER
+
+export interface PlayerAction {
+  type: ValidPlayerReducerActionTypes
+  payload: { id: number; name: string }
+}
